feat(TUP-29014): bugfix 2021-03-08

+++ main/plugins/org.talend.designer.runprocess/src/main/java/org/talend/designer/runprocess/java/TalendJavaProjectManager.java
@@ -433,13 +433,16 @@ public class TalendJavaProjectManager {
         try {
             String projectId = getCodesJarProjectId(type, projectTechName, codesJarName);
             ITalendProcessJavaProject project = talendCodesJarJavaProjects.get(projectId);
-            if (project != null) {
+
+            if (project != null && project.exists()) {
                 project.getProject().delete(deleteContent, true, null);
             }
             if (deleteContent) {
                 IFolder folder = new AggregatorPomsHelper(projectTechName).getCodeFolder(type).getFolder(codesJarName);
                 AggregatorPomsHelper.removeFromParentModules(folder.getFile(TalendMavenConstants.POM_FILE_NAME));
-                folder.delete(true, false, null);
+                if (folder.exists()) {
+                    folder.delete(true, false, null);
+                }
             }
             talendCodesJarJavaProjects.remove(projectId);
         } catch (Exception e) {
@@ -458,7 +461,7 @@ public class TalendJavaProjectManager {
                 Property property = project.getPropery();
                 if (property != null) {
                     IPath jobPath = ItemResourceUtil.getItemRelativePath(property);
-                    if (folderPath.isPrefixOf(jobPath)) {
+                    if (folderPath.isPrefixOf(jobPath) && project.exists()) {
                         project.getProject().delete(deleteContent, true, null);
                         iterator.remove();
                     }
@@ -468,7 +471,9 @@ public class TalendJavaProjectManager {
                 // delete folder
                 AggregatorPomsHelper helper = new AggregatorPomsHelper();
                 IFolder folder = helper.getProcessFolder(processType).getFolder(folderPath);
-                folder.delete(true, false, null);
+                if (folder.exists()) {
+                    folder.delete(true, false, null);
+                }
             }
         } catch (CoreException e) {
             ExceptionHandler.process(e);
@@ -507,7 +512,9 @@ public class TalendJavaProjectManager {
             String key = iterator.next();
             if (key.contains(id)) {
                 ITalendProcessJavaProject projectToDelete = talendJobJavaProjects.get(key);
-                projectToDelete.getProject().delete(deleteContent, true, null);
+                if (projectToDelete.exists()) {
+                    projectToDelete.getProject().delete(deleteContent, true, null);
+                }
                 String version = key.split("\\|")[1]; //$NON-NLS-1$
                 removedVersions.add(version);
                 iterator.remove();
